removed all member variables from bit_array_base while keeping dpf::bitstring trivially copyable
this was easier said than done, as types with virtual inheritance are not trivially copyable and so the base class needs some sneaky way to access methods from the child *without relying on inheritance* and *without storing a pointer to the child*.

Currently, bitstring uses const_cast to return a non-const pointer to its memory in data() const. That's not good. Also, setbit_iterable and parallel_bit_iterable need to be updated to match